Move some test from cover/ to nocover/
