+++ hypothesis-python/tests/cover/test_health_checks.py
@@ -13,13 +13,10 @@ import time
 import pytest
 from pytest import raises
 
-from hypothesis import HealthCheck, Phase, given, settings, strategies as st
+from hypothesis import HealthCheck, given, settings, strategies as st
 from hypothesis.control import assume
 from hypothesis.errors import FailedHealthCheck, InvalidArgument
 from hypothesis.internal.compat import int_from_bytes
-from hypothesis.internal.conjecture.data import ConjectureData
-from hypothesis.internal.conjecture.engine import BUFFER_SIZE
-from hypothesis.internal.entropy import deterministic_PRNG
 from hypothesis.stateful import (
     RuleBasedStateMachine,
     initialize,
@@ -27,7 +24,6 @@ from hypothesis.stateful import (
     rule,
     run_state_machine_as_test,
 )
-from hypothesis.strategies._internal.lazy import LazyStrategy
 from hypothesis.strategies._internal.strategies import SearchStrategy
 
 from tests.common.utils import no_shrink
@@ -122,21 +118,6 @@ def test_filtering_most_things_fails_a_health_check():
     assert "filter" in e.value.args[0]
 
 
-large_strategy = st.binary(min_size=7000, max_size=7000)
-too_large_strategy = st.tuples(large_strategy, large_strategy)
-
-
-def test_large_data_will_fail_a_health_check():
-    @given(st.none() | too_large_strategy)
-    @settings(database=None)
-    def test(x):
-        pass
-
-    with raises(FailedHealthCheck) as e:
-        test()
-    assert "allowable size" in e.value.args[0]
-
-
 def test_returning_non_none_is_forbidden():
     @given(st.integers())
     def a(x):
@@ -164,28 +145,6 @@ def test_the_slow_test_health_only_runs_if_health_checks_are_on():
     a()
 
 
-def test_large_base_example_fails_health_check():
-    @given(large_strategy)
-    def test(b):
-        pass
-
-    with pytest.raises(FailedHealthCheck) as exc:
-        test()
-
-    assert str(HealthCheck.large_base_example) in str(exc.value)
-
-
-def test_example_that_shrinks_to_overrun_fails_health_check():
-    @given(too_large_strategy | st.none())
-    def test(b):
-        pass
-
-    with pytest.raises(FailedHealthCheck) as exc:
-        test()
-
-    assert str(HealthCheck.large_base_example) in str(exc.value)
-
-
 class sample_test_runner:
     @given(st.none())
     def test(self, _):
@@ -208,72 +167,6 @@ def test_it_is_an_error_to_suppress_non_iterables():
 def test_it_is_an_error_to_suppress_non_healthchecks():
     with raises(InvalidArgument):
         settings(suppress_health_check=[1])
-
-
-slow_down_init = True
-
-
-def slow_init_integers(*args, **kwargs):
-    # This mimics st.characters() or st.text(), which perform some
-    # expensive Unicode calculations when the cache is empty.
-    global slow_down_init
-    if slow_down_init:
-        time.sleep(0.5)  # We monkeypatch time, so this is fast
-        slow_down_init = False
-    return st.integers(*args, **kwargs)
-
-
-@given(st.data())
-def test_lazy_slow_initialization_issue_2108_regression(data):
-    # Slow init in strategies wrapped in a LazyStrategy, inside an interactive draw,
-    # should be attributed to drawing from the strategy (not the test function).
-    # Specifically, this used to fail with a DeadlineExceeded error.
-    data.draw(LazyStrategy(slow_init_integers, (), {}))
-
-
-def test_does_not_trigger_health_check_on_simple_strategies(monkeypatch):
-    existing_draw_bits = ConjectureData.draw_bits
-
-    # We need to make drawing data artificially slow in order to trigger this
-    # effect. This isn't actually slow because time is fake in our CI, but
-    # we need it to pretend to be.
-    def draw_bits(self, n, *, forced=None, fake_forced=False):
-        time.sleep(0.001)
-        return existing_draw_bits(self, n, forced=forced, fake_forced=fake_forced)
-
-    monkeypatch.setattr(ConjectureData, "draw_bits", draw_bits)
-
-    with deterministic_PRNG():
-        for _ in range(100):
-            # Setting max_examples=11 ensures we have enough examples for the
-            # health checks to finish running, but cuts the generation short
-            # after that point to allow this test to run in reasonable time.
-            @settings(database=None, max_examples=11, phases=[Phase.generate])
-            @given(st.binary())
-            def test(b):
-                pass
-
-            test()
-
-
-def test_does_not_trigger_health_check_when_most_examples_are_small(monkeypatch):
-    with deterministic_PRNG():
-        for _ in range(100):
-            # Setting max_examples=11 ensures we have enough examples for the
-            # health checks to finish running, but cuts the generation short
-            # after that point to allow this test to run in reasonable time.
-            @settings(database=None, max_examples=11, phases=[Phase.generate])
-            @given(
-                st.integers(0, 100).flatmap(
-                    lambda n: st.binary(
-                        min_size=min(n * 100, BUFFER_SIZE), max_size=n * 100
-                    )
-                )
-            )
-            def test(b):
-                pass
-
-            test()
 
 
 class ReturningRuleMachine(RuleBasedStateMachine):
